% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{biweight}
\alias{biweight}
\title{Biweight kernel}
\usage{
biweight(u)
}
\arguments{
\item{u}{a vector of points at which to evaluate the kernel.}
}
\value{
a vector giving the evaluations.
}
\description{
Evaluates the biweight kernel, \deqn{K(u)=\frac{15}{16} (1-u^2)^2 I[|u|\leq 1].} This is a highly efficient kernel which is continuous and has a continuous first derivative.
}
\seealso{
Other kernel functions: 
\code{\link{epanechnikov}()},
\code{\link{gauss}()},
\code{\link{uniform}()}
}
\concept{kernel functions}
