% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch1data}
\alias{ch1data}
\title{Simulated data for Chapter 1}
\format{
A data frame with 200 rows and 2 columns:
\describe{
 \item{x}{design points}
 \item{y}{response values}
}
}
\usage{
ch1data
}
\description{
The simulated dataset used in Chapter 1. The design points are generated as independent samples from U(0,1).
The responses were generated as noisy observations from the function \deqn{m(x)=(x^2+1)sin(2\pi x(1+3x)) .}
The noise was simulated from a \eqn{N(0,0.1^2)}
}
\keyword{datasets}
