% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epanechnikov}
\alias{epanechnikov}
\title{Epanenchnikov kernel}
\usage{
epanechnikov(u)
}
\arguments{
\item{u}{a vector of points at which to evaluate the kernel.}
}
\value{
a vector giving the evaluations.
}
\description{
Evaluates the Epanechnikov kernel, \deqn{K(u)=\frac{3}{4} (1-u^2) I[|u|\leq 1].} This is the asymptotically optimal second order kernel. It is continuous, but the first derivative is discontinuous at \eqn{u=-1,1}.
}
\seealso{
Other kernel functions: 
\code{\link{biweight}()},
\code{\link{gauss}()},
\code{\link{uniform}()}
}
\concept{kernel functions}
