% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{gauss}
\alias{gauss}
\title{Gaussian kernel}
\usage{
gauss(u)
}
\arguments{
\item{u}{a vector of points at which to evaluate the kernel.}
}
\value{
a vector giving the evaluations.
}
\description{
Evaluates the Gaussian kernel, equal to the standard normal density. This is an alias for \code{dnorm}.
}
\seealso{
Other kernel functions: 
\code{\link{biweight}()},
\code{\link{epanechnikov}()},
\code{\link{uniform}()}
}
\concept{kernel functions}
