% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Compute exact MSE of a kernel estimator}
\usage{
mse(
  x_data,
  h,
  m,
  s2,
  estimator,
  t = seq(from = min(x_data), to = max(x_data), length = 100),
  ...
)
}
\arguments{
\item{x_data}{a vector giving the \eqn{x}-values in the data set}

\item{h}{a scalar, giving the bandwidth}

\item{m}{an R function, corresponding to the true regression function}

\item{s2}{a scalar giving the true response variance}

\item{estimator}{the estimator used (\code{nw}, \code{local_average}, or another estimator with the same input and output type)}

\item{t}{(optional) a vector specifying the points at which to evaluate the estimator. The default is a sequence of length 100 spanning the range of the \eqn{x}-data.}

\item{...}{additional arguments to be passed to \code{estimator} (e.g. \code{kernel} or \code{empty_nhood}).}
}
\value{
A list with 4 components:
\item{t}{a vector containing the evaluation points used}
\item{bias}{a vector containing the bias of the estimator at the evaluation points}
\item{var}{a vector containing the variance of the estimator at the evaluation points}
\item{mse}{a vector containing the mean-squared error of the estimator at the evaluation points}
}
\description{
This function computes the exact conditional bias, variance, and MSE of a kernel estimator. 'Conditional' means conditional given the design.
The estimator must be a linear smoother from this package.
The true regression function and response variance must both be known, therefore this function is only useful for theoretical examples.
}
\examples{
m <- function(x) { (x^2+1)*sin(2*pi*x*((1-x) + 4*x)) }
x <- sort(runif(80))
res <- mse(x, 0.02, m, 0.1^2, nw)
par(mfrow=c(1,2))
plot(res$t, res$mse, type="l", xlab="x", ylab="MSE")
plot(res$t, res$var, type="l", xlab="x", ylab="Variance")
rug(x)

}
