% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.npfit}
\alias{plot.npfit}
\alias{lines.npfit}
\title{Plotting functions for npfit objects}
\usage{
\method{plot}{npfit}(x, ...)

\method{lines}{npfit}(x, ...)
}
\arguments{
\item{x}{an \code{npfit} object, the result of running \code{nw} or \code{local_average}}

\item{...}{additional arguments to be passed to the plotting function, e.g. \code{xlim}, \code{col}, \code{lty} etc.}
}
\description{
Plotting methods for fitted nonparametric regression objects.
The \code{plot} command creates a new plot, with the observed data as points and the fitted nonparametric estimator as a line.
The \code{lines} command adds the nonparametric estimate to an existing plot, as a line.
}
\examples{
 m <- function(x) (x^2+1)*sin(2*pi*x*((1-x) + 4*x))
 x <- sort(runif(100))
 y <- m(x) + rnorm(length(x), sd=0.1)
 simdata <- data.frame(x=x,y=y)
 plot(local_average(simdata,h=0.02))
 lines(local_average(simdata,h=0.04), col=4)

}
