% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.npfit}
\alias{print.npfit}
\title{Print npfit object}
\usage{
\method{print}{npfit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{npfit}}

\item{...}{additional arguments (not needed)}
}
\description{
Pretty printing for nonparametric fits from \code{nw} or \code{local_average} etc.
}
