% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{uniform}
\alias{uniform}
\title{Uniform kernel}
\usage{
uniform(u)
}
\arguments{
\item{u}{a vector of points at which to evaluate the kernel.}
}
\value{
a vector giving the evaluations.
}
\description{
Evaluates the uniform kernel, \deqn{K(u)=\frac{1}{2} I[|u|\leq 1].}
}
\seealso{
Other kernel functions: 
\code{\link{biweight}()},
\code{\link{epanechnikov}()},
\code{\link{gauss}()}
}
\concept{kernel functions}
